Remove musikr
Remove musikr from my profile. I've lost interest in the project and I honestly don't think it's practical to work on two "Perpetual projects" at the same time. It's also not very pragmatic since the library is both unfinished and focused on such a niche subject that unless I were to make it public I doubt it would be that useful. In the future I may start working on it again, but currently Auxio and school projects are what I'm working on.